i transfered the json location to storage/private  and more...
removed text limit for business registration number and tin number since it is saved as encrypted.

change the location for the json files for the location and business type to storage/private and made sure it wont be ignored by git

change the hierarchy of some migration files since it was causing confliction.

temporarily turned item it foreign key to comment since it was causing conflict.